test(FlywheelSystem): Adapt to new paradigm
